Added function so that changes in data and ao3 goes directly to the export db as well.

+++ core-bc/composites/svc/src/main/java/se/vgregion/arbetsplatskoder/export/repository/ViewApkForSesamLmnRepository.java
@@ -8,6 +8,6 @@ import se.vgregion.arbetsplatskoder.export.repository.extension.Viewapkwithao3Ex
 /**
  * @author clalu4
  */
-public interface ViewApkForSesamLmnRepository extends JpaRepository<Viewapkforsesamlmn, Integer> {
+public interface ViewApkForSesamLmnRepository extends JpaRepository<Viewapkforsesamlmn, Long> {
 
 }
